requirements.txt
Django==3.2.3
tqdm==4.61.2
deepface==0.0.65
django_versatileimagefield==2.1
face_recognition==1.3.0
requests==2.25.1
djangorestframework==3.12.4
Pillow==8.3.1
